requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0
sqlalchemy>=2.0
pandas>=2.0
matplotlib>=3.8
python-dotenv>=1.0
tabulate>=0.9
streamlit